+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add iDEAL webhook support
 * Add AuthorizationAdjustment class and `authorization_adjustments` to Transaction
 * Coinbase is no longer a supported payment method. `PaymentMethodNoLongerSupported` will be returned for Coinbase operations.
+* Add facilitated details to Transaction if present
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`